some changes

+++ src/main/java/com/darts/server/repository/HospitalRepository.java
@@ -3,7 +3,7 @@ package com.darts.server.repository;
 import org.springframework.data.repository.CrudRepository;
 import com.darts.server.model.Hospital;
 
-public interface HospitalRepository extends CrudRepository<Hospital,Long>{
+public interface HospitalRepository extends CrudRepository<Hospital,Integer>{
     
 
 }
